refactor: renamed confusing `di.rs` into `env_impl.rs` to match near-sdk + enabled doctest
